cart put, delete method fix

+++ hyunphoto/cart/views.py
@@ -1,3 +1,4 @@
+from django.http import JsonResponse
 from django.shortcuts import get_object_or_404, render, redirect
 from rest_framework import status
 from rest_framework.views import APIView
@@ -27,7 +28,6 @@ class CartView(APIView):
 
     def get(self, request, *args, **kwargs):
         cart = Cart.objects.filter(user=request.user)
-
         total = self.get_total(self.request)
         context = {
             'user': request.user,
@@ -35,40 +35,28 @@ class CartView(APIView):
             'total': total,
         }
         return render(request, 'cart.html', context)
-    
-    # def post(self, request, *args, **kwargs):
-    #     action_type = request.data.get('action_type')
-    #     cart_item_id = request.data.get('cart_item')
-
-    #     if action_type == 'delete':
-    #         cart_item = get_object_or_404(Cart, id=cart_item_id, user=request.user)
-    #         cart_item.delete()
-    #         return redirect('cart')
-        
-    #     # elif action_type == 'modify':
-    #     #     new_quantity = request.POST.get('quantity', 1)
-    #     #     Cart.objects.filter(id=cart_item_id, user=request.user).update(quantity=new_quantity)
-    #     #     return redirect('cart')
-        
-    #     return Response(status=status.HTTP_400_BAD_REQUEST)
 
     def put(self, request, *args, **kwargs):
         cart_item_id = request.data.get('cart_item')
         new_quantity = request.data.get('quantity')
-        print('cart_item_id, new_quantity', cart_item_id, new_quantity)
+        print('cart_item_id:', cart_item_id, 'new_quantity:', new_quantity)
         if not cart_item_id or not new_quantity:
             return Response({'error': 'Invalid data'}, status=status.HTTP_400_BAD_REQUEST)
         
         Cart.objects.filter(id=cart_item_id).update(quantity = new_quantity)
-        return redirect('cart')
+        return JsonResponse({'satus': 'success'}, status=status.HTTP_200_OK)
     
     def delete(self, request, *args, **kwargs):
         cart_item_id = request.data.get('cart_item')
+        print('delete', cart_item_id)
         if not cart_item_id:
             return Response({'error': 'Invalid data'}, status=status.HTTP_400_BAD_REQUEST)
 
-        cart_item = get_object_or_404(Cart, id=cart_item_id, user=request.user)
-        cart_item.delete()
-        return redirect('cart')
+        # cart_item = get_object_or_404(Cart, id=cart_item_id, user=request.user)
+        cart_item = Cart.objects.filter(user=request.user, id=cart_item_id)
+        if cart_item:
+            cart_item.delete()
+            return JsonResponse({'satus': 'success'}, status=status.HTTP_200_OK)
+        return Response(status=status.HTTP_404_NOT_FOUND)
     
 cartview = CartView.as_view()
